과제끗

+++ 조명/실습34/Run_time_Framework.h
@@ -22,8 +22,8 @@ struct Ground {
 };
 struct Light {
 	GLfloat position[4] = { 0.0,0.0,0.0,1.0 };
-	GLfloat AmbientColor[4] = { 0.2,0.2,0.2,1.0 };
-	GLfloat DiffuseColor[4] = { 0.2,0.2,0.2,1.0 };
+	GLfloat AmbientColor[4] = { (float)0.2,(float)0.2,(float)0.2,(float)1.0 };
+	GLfloat DiffuseColor[4] = { (float)0.2,(float)0.2,(float)0.2,1.0 };
 	GLfloat SpecularColor[4] = { 1.0,1.0,1.0,1.0 };
 	GLfloat degree = 0.0;
 	bool on = true;
